requirements.txt
discord.py
asyncio
aiohttp
discord-webhook